+++ Changelog.md
@@ -1,5 +1,13 @@
 # Better Inventory changelog
 
+## v0.9.1.2
+- Fixed Return to Previous Slot not working as expected when dropping favorited items
+- Fixed Void Bag First not handling coins properly
+- Fixed RecipeUI consuming keyboard input even when not visible
+- Fixed Quick Stack hotkey not refreshing recipes
+- Fixed Quick Stack on Pickup not handling multiplayer properly
+- Fixed Quick Stack on Pickup not working properly when a chest was opened
+
 ## v0.9.1.1
 - Fixed Quick Stack on item pickup deleting coins